Performance block

+++ apple/objc/interface/monitor/WCTDatabase+Monitor.h
@@ -30,7 +30,7 @@ typedef void (^WCTErrorTraceBlock)(WCTError *);
 /**
  Trigger when a transaction or a normal sql ends.
  */
-typedef void (^WCTPerformanceTraceBlock)(NSArray<WCTPerformanceFootprint *> *, NSInteger /* cost */);
+typedef void (^WCTPerformanceTraceBlock)(NSArray<WCTPerformanceFootprint *> *, double /* cost */);
 
 /**
  Trigger when a SQL is executed.
